date int overflow

+++ fe/fe-core/src/main/java/com/starrocks/connector/iceberg/cost/StatUtils.java
@@ -41,11 +41,11 @@ public class StatUtils {
         } else if (type instanceof Types.TimestampType) {
             // we deal iceberg TimestampType as seconds in columnstatistics
             // in iceberg it's microsecond
-            valueConvert = (long) value / 1000000;
+            valueConvert = ((long) value) / 1000000;
         } else if (type instanceof Types.DateType) {
             // we deal iceberg DateType as seconds in columnstatistics
             // in iceberg it's num of day from 1970-01-01
-            valueConvert = (int) value * 86400;
+            valueConvert = ((long) ((int) value)) * 86400;
         } else if (type instanceof Types.DecimalType) {
             valueConvert = ((BigDecimal) value).doubleValue();
         } else {
